'finalised' id/fd but I'm discovering some problems. The torque actuator stuff is making the arm windmill. I realize that we may know the torques in the arm in its static position, but this might not necessarily translate to the torques in the virtual moving position. I'm also wondering if a vec2 containing udot information is enough to properly portray the state - that is, how can we move laterally if we only have a number that represents torque around one axis?
further, and less urgently, I'm not sure if the controls for the actuators in ID/FD are realistic. If the control output is 1, is the force then 1N, or is the force the optimal force? and so we might need to scale the control input. But that's a //TODO for another day