added logic to perform FD directly without relying on ID torques. This isn't working, however. I think I may have similar problems to what happened with the ID model before, but I'm not sure what that is. Some thought required.
Also a lot of debugging

I'm not certain that this is the most accurate approach, but it might be the easiest. More research required to find out how I can add in the ID torques to my FD. The problem is that the torques in isometric are always going to do weird things in free motion, since a static arm will provide torque over a long period without moving, and in the meantime the virtual arm has spun around the world from that torque. I guess that leaves the responsibility with the subject to adapt, providing less torque at a time and reversing direction...